+++ selenium-csharp/src/SeleniumCSharp/Options/Credentials.cs
@@ -1,0 +1,7 @@
+namespace SeleniumCSharp.Options;
+
+public class Credentials
+{
+    public string EmailAddress { get; set; } = default!;
+    public string Password { get; set; } = default!;
+}
