+++ selenium-csharp/src/SeleniumCSharp/Options/PublicProfile.cs
@@ -1,0 +1,7 @@
+namespace SeleniumCSharp.Options;
+
+public class PublicProfile
+{
+    public string Username { get; set; } = default!;
+    public string ExpectedFullname { get; set; } = default!;
+}
